Rename Redux module: signIn => auth
