+++ src/frontend/stories/Layout/component.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import {FacebookProfile} from '~/../utils/models/FacebookProfile'
 
 import * as api        from '~/frontend/api'
-import {IActions}      from '~/frontend/redux/actions/signIn'
+import {IActions}      from '~/frontend/redux/actions/auth'
 import LayoutComponent from '~/frontend/components/Layout'
 
 
@@ -15,7 +15,7 @@ interface ILayout {
 class Layout extends React.Component<ILayout, {}> {
   componentWillMount () {
     const setFacebookId = (profile: FacebookProfile) => {
-      this.props.actions.setFacebookId(profile.id)
+      this.props.actions.signIn(profile.id)
     }
     api.getProfile()
       .then(setFacebookId)
